SpreadsheetCellReferenceOrLabelName.viewport pulled from SpreadsheetCellReference
